Clean up and refactor package placement for util related classes
