+++ core-lib/src/main/java/org/idpass/smartscanner/lib/SmartScannerActivity.kt
@@ -67,15 +67,15 @@ import org.idpass.smartscanner.lib.idpasslite.IDPassManager
 import org.idpass.smartscanner.lib.mrz.MRZAnalyzer
 import org.idpass.smartscanner.lib.mrz.MrzUtils
 import org.idpass.smartscanner.lib.nfc.NFCScanAnalyzer
-import org.idpass.smartscanner.lib.platform.BaseActivity
-import org.idpass.smartscanner.lib.platform.extension.*
-import org.idpass.smartscanner.lib.platform.utils.CameraUtils.isLedFlashAvailable
-import org.idpass.smartscanner.lib.platform.utils.LanguageUtils
 import org.idpass.smartscanner.lib.platform.utils.PlayStoreUtils
-import org.idpass.smartscanner.lib.platform.utils.transform.CropTransformation
+import org.idpass.smartscanner.lib.scanner.BaseActivity
 import org.idpass.smartscanner.lib.scanner.ImageResult
 import org.idpass.smartscanner.lib.scanner.SmartScannerException
 import org.idpass.smartscanner.lib.scanner.config.*
+import org.idpass.smartscanner.lib.utils.CameraUtils.isLedFlashAvailable
+import org.idpass.smartscanner.lib.utils.LanguageUtils
+import org.idpass.smartscanner.lib.utils.extension.*
+import org.idpass.smartscanner.lib.utils.transform.CropTransformation
 import java.io.File
 import java.util.*
 import java.util.concurrent.ExecutorService
