+++ core-lib/src/main/java/org/idpass/smartscanner/lib/idpasslite/IDPassManager.kt
@@ -32,8 +32,8 @@ import org.idpass.lite.exceptions.InvalidCardException
 import org.idpass.lite.exceptions.InvalidKeyException
 import org.idpass.smartscanner.api.ScannerConstants
 import org.idpass.smartscanner.lib.SmartScannerActivity
-import org.idpass.smartscanner.lib.platform.utils.DateUtils
 import org.idpass.smartscanner.lib.scanner.config.Modes
+import org.idpass.smartscanner.lib.utils.DateUtils
 
 
 object IDPassManager {
