+++ core-lib/src/main/java/org/idpass/smartscanner/lib/nfc/NFCFragment.kt
@@ -39,11 +39,11 @@ import org.idpass.smartscanner.lib.R
 import org.idpass.smartscanner.lib.nfc.details.IntentData
 import org.idpass.smartscanner.lib.nfc.details.NFCDocumentTag
 import org.idpass.smartscanner.lib.nfc.passport.Passport
-import org.idpass.smartscanner.lib.platform.utils.DateUtils
-import org.idpass.smartscanner.lib.platform.utils.DateUtils.formatStandardDate
-import org.idpass.smartscanner.lib.platform.utils.KeyStoreUtils
-import org.idpass.smartscanner.lib.platform.utils.LanguageUtils
 import org.idpass.smartscanner.lib.scanner.config.Language
+import org.idpass.smartscanner.lib.utils.DateUtils
+import org.idpass.smartscanner.lib.utils.DateUtils.formatStandardDate
+import org.idpass.smartscanner.lib.utils.KeyStoreUtils
+import org.idpass.smartscanner.lib.utils.LanguageUtils
 import org.jmrtd.*
 import org.jmrtd.lds.icao.MRZInfo
 import org.spongycastle.jce.provider.BouncyCastleProvider
