+++ core-lib/src/main/java/org/idpass/smartscanner/lib/scanner/BaseActivity.kt
@@ -15,14 +15,16 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.platform.extension
+package org.idpass.smartscanner.lib.scanner
 
-import android.app.Activity
-import android.content.Context
-import android.view.View
-import android.view.inputmethod.InputMethodManager
+import androidx.appcompat.app.AppCompatActivity
 
-fun Context.hideKeyboard(view: View) {
-    val inputMethodManager = getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
-    inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
+abstract class BaseActivity : AppCompatActivity() {
+
+    fun hideActionBar() {
+        supportActionBar?.hide()
+        supportActionBar?.setDisplayShowTitleEnabled(false)
+        actionBar?.hide()
+        actionBar?.setDisplayShowTitleEnabled(false)
+    }
 }
