+++ core-lib/src/main/java/org/idpass/smartscanner/lib/utils/DateUtils.kt
@@ -15,7 +15,7 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.platform.utils
+package org.idpass.smartscanner.lib.utils
 
 import android.annotation.SuppressLint
 import org.idpass.smartscanner.lib.scanner.config.Language
