+++ core-lib/src/main/java/org/idpass/smartscanner/lib/utils/ImageUtils.kt
@@ -15,7 +15,7 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.platform.utils
+package org.idpass.smartscanner.lib.utils
 
 import android.graphics.*
 import android.media.Image
