+++ core-lib/src/main/java/org/idpass/smartscanner/lib/utils/PlayStoreUtils.kt
