+++ core-lib/src/main/java/org/idpass/smartscanner/lib/utils/extension/Context.kt
@@ -15,9 +15,14 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.platform.extension
+package org.idpass.smartscanner.lib.utils.extension
 
+import android.app.Activity
+import android.content.Context
+import android.view.View
+import android.view.inputmethod.InputMethodManager
 
-fun Double.Companion.noValue() = 0.0
-
-fun Int.Companion.noValue() = 0
+fun Context.hideKeyboard(view: View) {
+    val inputMethodManager = getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
+    inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
+}
