+++ core-lib/src/main/java/org/idpass/smartscanner/lib/utils/transform/BitmapTransformation.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.idpass.smartscanner.lib.platform.utils.transform;
+package org.idpass.smartscanner.lib.utils.transform;
 
 import android.content.Context;
 import android.graphics.Bitmap;
